react: Allow overriding <DashboardModal /> `target` prop
This was always mounting the dashboard DOM element into the current
React container element, but that's not always the right thing to do,
for example if the container element is a small element with `overflow:
hidden` or some `position: relative` style.

This patch allows you to do `target={document.body}` and the like, so
you can mount the dashboard element anywhere.