Add object rest spread transform
Object rest spread was standardised this year so it's safe to use. It
has shipped in major browsers and Node. Because Uppy uses immutable
state we can make a lot of gains in terseness and readability by using
the new object syntaxes, I think.

I updated one place where we were using Object.assign; I think we can
merge it like this first and then update the other places when we are
changing stuff there anyway.