Remove locales
Fixes #1348

Locales have been outdated for some time, while we were planning on updating everything. The docs for each plugin https://uppy.io/docs/drag-drop/#locale and source code are the source of truth right now.

The reason the folder is around, is we thought we’ll use some of the old strings to help translate the news ones. Can do that using history, so removing to avoid confusion.