tus: docs-deprecate autoRetry (#2347)
* tus: deprecate autoRetry

`autoRetry` retries failed uploads when the user's internet connection
comes back up after an outage. It's not totally clear from the name that
that is what it does: it could also just retry automatically after a
failure generally.

This feature can be implemented by users like this:
```js
uppy.on('back-online', () => {
  uppy.retryAll()
})
```
I think that's clearer than `autoRetry: true`, and we don't even have
to come up with a name!

Since the implementation is trivial and the functionality is independent
of tus, it probably shouldn't live in the plugin. Users can copy the
three lines from the docs and use it that way. We could also consider
moving it into `@uppy/core`.

* tus: fill in pr url

* tus: revert runtime deprecation

* changelog: remove deprecated `autoRetry` in 2.0